GH-46121: [Python] Add missing `column_index` argument to `ArrowReaderProperties::read_dictionary`'s Cython binding (#46122)
### Rationale for this change
This PR adds the missing `column_index` argument to `ArrowReaderProperties::read_dictionary`'s Cython binding to resolve the issue described in https://github.com/apache/arrow/issues/46121.

### What changes are included in this PR?
1. Add the missing `column_index` argument to `ArrowReaderProperties::read_dictionary`'s Cython binding.

### Are these changes tested?
No (I'm not sure where the tests for Cython bindings are; I'm happy to add a test if someone can point me in the right direction)

### Are there any user-facing changes?
Yes; this is a bug fix to a user facing API (though it's admittedly unlikely to be used).
* GitHub Issue: #46121

Authored-by: Lester Fan <lesterfan37@gmail.com>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>