GH-46908: [Docs][Format] Add variant extension type docs (#47456)
### Rationale for this change
To support the addition of the Parquet Variant data type and the Iceberg
adoption of the variant type, we need a defined way to pass this data
through Arrow-compatible systems. As such, we need a specification for a
canonical Arrow extension type to represent Variant data.

### What changes are included in this PR?
Updates to the docs which define the Arrow Variant Extension type


* GitHub Issue: #46908

---------

Co-authored-by: Ian Cook <ianmcook@gmail.com>
Co-authored-by: David Li <li.davidm96@gmail.com>
Co-authored-by: Dewey Dunnington <dewey@dunnington.ca>
Co-authored-by: Gang Wu <ustcwg@gmail.com>
Co-authored-by: Antoine Pitrou <pitrou@free.fr>
Co-authored-by: Sutou Kouhei <kou@cozmixng.org>
Co-authored-by: Bryce Mecum <petridish@gmail.com>
Co-authored-by: Yan Tingwang <tingwangyan2020@163.com>