GH-42137 [CI][Python] Add Python Windows GitHub Action and remove AppVeyor (#47567)
### Rationale for this change

There is currently a single CI job on AppVeyor. This is the only job on Windows for Python and it usually fails due to CI issues rather than Python real issues. We have more experience with GitHub actions and debugging AppVeyor is usually challenging.

### What changes are included in this PR?

Create a Python Windows job that runs on GitHub actions. The job builds Arrow CPP and PyArrow and it tests PyArrow.

Also removes the CI related files for AppVeyor

### Are these changes tested?

Yes, via CI.

### Are there any user-facing changes?

No

* GitHub Issue: #42137

Lead-authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Co-authored-by: Sutou Kouhei <kou@cozmixng.org>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>