GH-47751: [CI] Fix check for job to ignore on reporting (#47755)
### Rationale for this change

It failed notifying on Zulip and the email, see:
https://lists.apache.org/thread/12glcwk1jb8lb5x6lw3t2vr6zn9kp66v

We basically were ignoring all jobs except the report-ci :disappointed: 

### What changes are included in this PR?

The check was wrong, I clearly did not test this properly after the latest change, sorry about that.

### Are these changes tested?

I validated locally, without the change:
```
$ archery ci report-chat --dry-run --ignore report-ci --repository apache/arrow 18329667061

*Extra CI GitHub report for <https://github.com/apache/arrow/actions/runs/18329667061|C++ Extra>*

:tada: *1 successful jobs*
```
It shows successful because it has now finished, this is not the case when is running on CI, but only a single job which is not expected, it should show 5!

with the change:
```
archery ci report-chat --dry-run --ignore report-ci --repository apache/arrow 18329667061

*Extra CI GitHub report for <https://github.com/apache/arrow/actions/runs/18329667061|C++ Extra>*

:tada: *5 successful jobs*
```

### Are there any user-facing changes?

No

* GitHub Issue: #47751

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>