GH-47844: [CI] Fix unconditionally running extra workflows reporting when there are jobs failing (#47917)
### Rationale for this change

If a job fails on the workflow the report step is currently being skipped.

### What changes are included in this PR?

The unconditional `always()` check should be done on the parent workflow using `.github/workflows/report_ci.yml`  not on the common one.
Added also missing needed jobs before generating the report on the cpp_extra.yml (msvc-arm64 and jni-linux)

### Are these changes tested?

Yes via CI I validated the job will be executed even if there are job failures.

### Are there any user-facing changes?

No
* GitHub Issue: #47844

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: raulcd <raulcumplido@gmail.com>