GH-47778: [CI][Python] Remove ORC alias timezone for US/Pacific on test_orc.py::test_timezone_absent (#47956)
### Rationale for this change

Since we upgraded to ORC 2.1.1 our test started failing because it was able to resolved the timezone and read the file when we were expecting an exception to be raised.

In order to support Legacy timezones ORC added some aliases to some timezones, in the case of our test `US/Pacific` was aliased to `America/Los_Angeles` and the test was finding the timezone.

### What changes are included in this PR?

Remove both timezones `US/Pacific` and `America/Los_Angeles` so the test works as expected.

### Are these changes tested?

Yes, locally via CI and some extra archery tasks

### Are there any user-facing changes?

No

* GitHub Issue: #47778

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>